GPS Library Mod updated
Complies with code, more testing would be advised, however this should
be in a working state.

Modified the fullscale adaptation to use HardwareSerial on Serial 1 and
4 rather than software serial.

Modified the Adafruit library to forcefully remove references to
NewSoftwareSerial to avoid compatibility issues, as well as removed
UDR0? references, computation speed lost for functionality.

Corrections to Teensy hardware library files in order to correct this
adaptation.